Fix npm script
