+++ artifactory/services/fspatterns/utils.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"github.com/jfrog/jfrog-client-go/utils/log"
 	"os"
+	"path/filepath"
 	"regexp"
 	"strings"
 
@@ -25,8 +26,8 @@ func ListFilesFilterPatternAndSize(rootPath string, isRecursive, includeDirs, ex
 	if excludeWithRelativePath {
 		rootFilter = rootPath
 	}
-	filterFunc := filterFilesFunc(rootFilter, includeDirs, excludePathPattern, sizeThreshold)
-	return fileutils.ListFilesWithFilterFunc(rootPath, isRecursive, !isSymlink, filterFunc)
+	filterFunc := filterFilesFunc(rootFilter, isRecursive, includeDirs, excludePathPattern, sizeThreshold)
+	return fileutils.ListFilesWithFilterFunc(rootPath, !isSymlink, filterFunc)
 }
 
 // Transform to regexp and prepare Exclude patterns to be used, exclusion patterns must be absolute paths.
@@ -50,10 +51,16 @@ func PrepareExcludePathPattern(exclusions []string, patternType utils.PatternTyp
 }
 
 // Returns a function that filters files according to the provided parameters
-func filterFilesFunc(rootPath string, includeDirs bool, excludePathPattern string, sizeThreshold *SizeThreshold) func(filePath string) (included bool, err error) {
+func filterFilesFunc(rootPath string, isRecursive, includeDirs bool, excludePathPattern string, sizeThreshold *SizeThreshold) func(filePath string) (included bool, err error) {
 	return func(path string) (included bool, err error) {
 		if path == "." {
 			return false, nil
+		}
+		if !isRecursive {
+			isInRoot, err := isFileInRootDir(rootPath, path)
+			if err != nil || !isInRoot {
+				return false, err
+			}
 		}
 		if !includeDirs {
 			isDir, err := fileutils.IsDirExists(path, false)
@@ -83,6 +90,17 @@ func filterFilesFunc(rootPath string, includeDirs bool, excludePathPattern strin
 		}
 		return true, nil
 	}
+}
+
+func isFileInRootDir(rootPath, path string) (bool, error) {
+	relPath, err := filepath.Rel(rootPath, path)
+	if err != nil {
+		return false, err
+	}
+	if filepath.Dir(relPath) != "." {
+		return false, nil
+	}
+	return true, nil
 }
 
 // Return the actual sub-paths that match the regex provided.
